pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "简洁的个人博客"
authors = [{ name = "QING", email = "qlear@qq.com" }]
requires-python = ">=3.12,<4.0"
dependencies = [
    "flask<4.0.0,>=3.1.0", # web framework
    "python-frontmatter<2.0.0,>=1.1.0", # frontmatter parser
    "Flask-SQLAlchemy<4.0.0,>=3.1.1", # ORM extension
    "python-dotenv>=1.0.1",
    "loguru>=0.7.3",
    "gunicorn>=23.0.0",
    "flask-wtf>=1.2.2",
    "email-validator>=2.2.0",
    "flask-mailman>=1.1.1",
    "celery>=5.5.1",
    "redis>=5.2.1",
]

[dependency-groups]
dev = [
    "ruff<1.0.0,>=0.8.0",       # linter & formatter
    "pytest<8.4.0,>=8.3.4",     # test runner
    "pre-commit<5.0.0,>=4.0.1", # git hooks
]

[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
default = true

[tool.ruff]
exclude = [
    ".venv",
    ".idea",
    ".vscode",
    ".uv_cache",
    ".ruff_cache",
    ".pytest_cache",
    "__pycache__",
]
target-version = "py312"
line-length = 120
indent-width = 4

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG001", # unused arguments in functions
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
